add modlist check to avoid potential problems
