+++ src/main/java/com/robertx22/age_of_exile/database/data/spells/components/ProjectileCastHelper.java
@@ -5,7 +5,7 @@ import com.robertx22.age_of_exile.database.data.spells.entities.CalculatedSpellD
 import com.robertx22.age_of_exile.database.data.spells.map_fields.MapField;
 import com.robertx22.age_of_exile.database.data.spells.spell_classes.SpellCtx;
 import com.robertx22.age_of_exile.database.data.spells.spell_classes.SpellUtils;
-import com.robertx22.age_of_exile.event_hooks.player.OnLogin;
+import com.robertx22.age_of_exile.database.data.spells.spell_fx.FXInfoHolder;
 import com.robertx22.age_of_exile.uncommon.effectdatas.rework.EventData;
 import com.robertx22.age_of_exile.uncommon.utilityclasses.AllyOrEnemy;
 import com.robertx22.age_of_exile.uncommon.utilityclasses.EntityFinder;
@@ -112,7 +112,7 @@ public class ProjectileCastHelper {
                 Vec3 finalPosAdd = pos.add(posAdd);
                 getPlayerWithinRange(pos.add(posAdd), world, 128.0D)
                         .stream()
-                        .filter(OnLogin::readFXConfigValue)
+                        .filter(FXInfoHolder::readFXConfigValue)
                         .toList()
                         .forEach(serverPlayer ->
                                 Packets.sendToClient(serverPlayer, new SpellEntityInitPacket(en.getUUID(), new Vec3(finalPosAdd.x, caster.getEyeY() - 0.1F, finalPosAdd.z), holder.get(MapField.SKILL_FX))));
