+++ src/main/java/com/robertx22/age_of_exile/database/data/spells/entities/FXEntity.java
@@ -6,7 +6,7 @@ import com.robertx22.age_of_exile.database.data.spells.components.MapHolder;
 import com.robertx22.age_of_exile.database.data.spells.entities.renders.IMyRenderAsItem;
 import com.robertx22.age_of_exile.database.data.spells.map_fields.MapField;
 import com.robertx22.age_of_exile.database.data.spells.spell_classes.SpellCtx;
-import com.robertx22.age_of_exile.event_hooks.player.OnLogin;
+import com.robertx22.age_of_exile.database.data.spells.spell_fx.FXInfoHolder;
 import com.robertx22.age_of_exile.mmorpg.registers.common.SlashEntities;
 import com.robertx22.age_of_exile.uncommon.utilityclasses.Utilities;
 import com.robertx22.age_of_exile.vanilla_mc.packets.sendSpellEntityDeath;
@@ -123,7 +123,7 @@ public class FXEntity extends Entity implements IDatapackSpellEntity, IMyRenderA
         if(this.playerList.isEmpty()){
             this.playerList.addAll(getPlayerWithinRange(this.position(), this.level(), 128.0D)
                     .stream()
-                    .filter(OnLogin::readFXConfigValue)
+                    .filter(FXInfoHolder::readFXConfigValue)
                     .toList()
             );
         }
