+++ src/main/java/com/robertx22/age_of_exile/mmorpg/MMORPG.java
@@ -14,6 +14,7 @@ import com.robertx22.age_of_exile.database.data.loot_chest.base.LootChests;
 import com.robertx22.age_of_exile.database.data.profession.ProfessionEvents;
 import com.robertx22.age_of_exile.database.data.spells.components.conditions.EffectCondition;
 import com.robertx22.age_of_exile.database.data.spells.map_fields.MapField;
+import com.robertx22.age_of_exile.database.data.spells.spell_fx.FXInfoHolder;
 import com.robertx22.age_of_exile.database.data.stats.types.special.SpecialStats;
 import com.robertx22.age_of_exile.database.registrators.Currencies;
 import com.robertx22.age_of_exile.database.registry.ExileDBInit;
@@ -134,6 +135,7 @@ public class MMORPG {
 
         ModTags.init();
 
+        FXInfoHolder.init();
         StatEffects.loadClass();
         StatConditions.loadClass();
         Stats.loadClass();
