+++ src/main/java/com/robertx22/age_of_exile/vanilla_mc/packets/askForFXConfigPacket.java
@@ -1,16 +1,13 @@
 package com.robertx22.age_of_exile.vanilla_mc.packets;
 
-import com.robertx22.age_of_exile.config.forge.ClientConfigs;
 import com.robertx22.age_of_exile.mmorpg.SlashRef;
 import com.robertx22.library_of_exile.main.MyPacket;
 import com.robertx22.library_of_exile.main.Packets;
 import com.robertx22.library_of_exile.packets.ExilePacketContext;
 import net.minecraft.network.FriendlyByteBuf;
-import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.server.packs.repository.Pack;
-
-import static com.robertx22.age_of_exile.event_hooks.player.OnLogin.writeFXConfigValueFromPacket;
+import net.minecraftforge.fml.ModList;
+import net.minecraftforge.forgespi.language.IModInfo;
 
 public class askForFXConfigPacket extends MyPacket<askForFXConfigPacket> {
 
@@ -35,11 +32,10 @@ public class askForFXConfigPacket extends MyPacket<askForFXConfigPacket> {
 
     @Override
     public void onReceived(ExilePacketContext ctx) {
-        ctx.getPlayer().displayClientMessage(Component.literal("get ask!"), false);
         var callback = new FXConfigCheckerPacket();
-        Packets.sendToServer(callback);
-        ctx.getPlayer().displayClientMessage(Component.literal("send back!"), false);
-
+        if(ModList.get().getMods().stream().map(IModInfo::getModId).anyMatch(s -> s.contentEquals("photon"))){
+            Packets.sendToServer(callback);
+        }
     }
 
     @Override
