+++ src/main/java/com/robertx22/age_of_exile/vanilla_mc/packets/sendSpellEntityPositionPacket.java
@@ -9,7 +9,7 @@ import net.minecraft.world.phys.Vec3;
 
 import java.util.UUID;
 
-import static com.robertx22.age_of_exile.event_hooks.player.OnLogin.clientPlayerEntityFXHolder;
+import static com.robertx22.age_of_exile.database.data.spells.spell_fx.FXInfoHolder.clientPlayerEntityFXHolder;
 
 public class sendSpellEntityPositionPacket extends MyPacket<sendSpellEntityPositionPacket> {
